[mathfx] Replacement of Segment by SegmentPoint.
