+++ advanced/mathfx/src/main/java/org/arakhne/afc/math/geometry/d2/dfx/Segment2dfx.java
@@ -23,7 +23,6 @@ package org.arakhne.afc.math.geometry.d2.dfx;
 import javafx.beans.binding.Bindings;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import org.eclipse.xtext.xbase.lib.Pure;
 
@@ -33,8 +32,11 @@ import org.arakhne.afc.math.geometry.d2.Transform2D;
 import org.arakhne.afc.math.geometry.d2.afp.Segment2afp;
 import org.arakhne.afc.vmutil.asserts.AssertMessages;
 
-/** Segment with 2 double precision floating-point FX properties.
+/** A 2D segment/line encapsulating points with 2 double precision FX properties.
  *
+ *  <p>This segment is defined by its two extremities. It should not differ from
+ *  the original Segment2dfx except from storage type.
+ * @author $Author: tpiotrow$
  * @author $Author: sgalland$
  * @author $Author: hjaffali$
  * @version $FullVersion$
@@ -43,17 +45,13 @@ import org.arakhne.afc.vmutil.asserts.AssertMessages;
  * @since 13.0
  */
 public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
-		implements Segment2afp<Shape2dfx<?>, Segment2dfx, PathElement2dfx, Point2dfx, Vector2dfx, Rectangle2dfx> {
+	implements Segment2afp<Shape2dfx<?>, Segment2dfx, PathElement2dfx, Point2dfx, Vector2dfx, Rectangle2dfx> {
 
 	private static final long serialVersionUID = -5603953934276693947L;
 
-	private DoubleProperty ax;
+	private Point2dfx p1 = new Point2dfx();
 
-	private DoubleProperty ay;
-
-	private DoubleProperty bx;
-
-	private DoubleProperty by;
+	private Point2dfx p2 = new Point2dfx();
 
 	/** Construct an empty segment.
 	 */
@@ -61,27 +59,51 @@ public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
 		//
 	}
 
-	/** Construct a segment with the two given points.
-	 * @param p1 first point.
-	 * @param p2 second point.
+	/** Constructor by setting.
+	 *
+	 * @param p1 the point to set as first point.
+	 * @param p2 the point to set as second point.
 	 */
+	public Segment2dfx(Point2dfx p1, Point2dfx p2) {
+	    assert p1 != null : AssertMessages.notNullParameter(0);
+	    assert p2 != null : AssertMessages.notNullParameter(1);
+	    this.p1 = p1;
+	    this.p2 = p2;
+	}
+
+	/** Construct a segment with the two given points.
+     * @param p1 first point.
+     * @param p2 second point.
+     */
 	public Segment2dfx(Point2D<?, ?> p1, Point2D<?, ?> p2) {
-		this(p1.getX(), p1.getY(), p2.getX(), p2.getY());
+	    assert p1 != null : AssertMessages.notNullParameter(0);
+	    assert p2 != null : AssertMessages.notNullParameter(1);
+		set(p1.getX(), p1.getY(), p2.getX(), p2.getY());
+	}
+
+	/** Constructor by setting.
+	 * @param segment the segment to set.
+	 */
+	public Segment2dfx(Segment2dfx segment) {
+	    assert segment != null : AssertMessages.notNullParameter();
+	    this.p1 = segment.p1;
+	    this.p2 = segment.p2;
 	}
 
 	/** Constructor by copy.
-	 * @param segment the segment to copy.
-	 */
+     * @param segment the segment to copy.
+     */
 	public Segment2dfx(Segment2afp<?, ?, ?, ?, ?, ?> segment) {
-		this(segment.getX1(), segment.getY1(), segment.getX2(), segment.getY2());
+	    assert segment != null : AssertMessages.notNullParameter();
+		set(segment.getX1(), segment.getY1(), segment.getX2(), segment.getY2());
 	}
 
 	/** Construct a segment with the two given points.
-	 * @param x1 x coordinate of the first point.
-	 * @param y1 y coordinate of the first point.
-	 * @param x2 x coordinate of the second point.
-	 * @param y2 y coordinate of the second point.
-	 */
+     * @param x1 x coordinate of the first point.
+     * @param y1 y coordinate of the first point.
+     * @param x2 x coordinate of the second point.
+     * @param y2 y coordinate of the second point.
+     */
 	public Segment2dfx(double x1, double y1, double x2, double y2) {
 		set(x1, y1, x2, y2);
 	}
@@ -89,21 +111,13 @@ public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
 	@Override
 	public Segment2dfx clone() {
 		final Segment2dfx clone = super.clone();
-		if (clone.ax != null) {
-			clone.ax = null;
-			clone.x1Property().set(getX1());
+		if (clone.p1 != null) {
+			clone.p1 = null;
+			clone.p1 = this.p1.clone();
 		}
-		if (clone.ay != null) {
-			clone.ay = null;
-			clone.y1Property().set(getY1());
-		}
-		if (clone.bx != null) {
-			clone.bx = null;
-			clone.x2Property().set(getX2());
-		}
-		if (clone.by != null) {
-			clone.by = null;
-			clone.y2Property().set(getY2());
+		if (clone.p2 != null) {
+			clone.p2 = null;
+			clone.p2 = this.p2.clone();
 		}
 		return clone;
 	}
@@ -111,25 +125,44 @@ public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
 	@Pure
 	@Override
 	public int hashCode() {
-		int bits = 1;
-		bits = 31 * bits + Double.hashCode(getX1());
-		bits = 31 * bits + Double.hashCode(getY1());
-		bits = 31 * bits + Double.hashCode(getX2());
-		bits = 31 * bits + Double.hashCode(getY2());
-        return bits ^ (bits >> 31);
+		long bits = 1;
+		bits = 31 * bits + Double.doubleToLongBits(getX1());
+		bits = 31 * bits + Double.doubleToLongBits(getY1());
+		bits = 31 * bits + Double.doubleToLongBits(getX2());
+		bits = 31 * bits + Double.doubleToLongBits(getY2());
+		final int b = (int) bits;
+		return b ^ (b >> 32);
+	}
+
+	@Pure
+	@Override
+	public String toString() {
+		final StringBuilder b = new StringBuilder();
+		b.append("["); //$NON-NLS-1$
+		b.append(getX1());
+		b.append(";"); //$NON-NLS-1$
+		b.append(getY1());
+		b.append("|"); //$NON-NLS-1$
+		b.append(getX2());
+		b.append(";"); //$NON-NLS-1$
+		b.append(getY2());
+		b.append("]"); //$NON-NLS-1$
+		return b.toString();
 	}
 
 	@Pure
 	@Override
 	public Segment2dfx createTransformedShape(Transform2D transform) {
-		assert transform != null : AssertMessages.notNullParameter();
+        if (transform == null || transform.isIdentity()) {
+            return clone();
+        }
 		final Point2dfx point = getGeomFactory().newPoint(getX1(), getY1());
 		transform.transform(point);
 		final double x1 = point.getX();
 		final double y1 = point.getY();
 		point.set(getX2(), getY2());
 		transform.transform(point);
-		return getGeomFactory().newSegment(x1, y1, point.getX(), point.getY());
+		return new Segment2dfx(x1, y1, point.getX(), point.getY());
 	}
 
 	@Override
@@ -163,7 +196,7 @@ public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
 	@Pure
 	@Override
 	public double getX1() {
-		return this.ax == null ? 0 : this.ax.get();
+		return this.p1.getX();
 	}
 
 	/** Replies the property that is the x coordinate of the first segment point.
@@ -172,16 +205,13 @@ public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
 	 */
 	@Pure
 	public DoubleProperty x1Property() {
-		if (this.ax == null) {
-			this.ax = new SimpleDoubleProperty(this, MathFXAttributeNames.X1);
-		}
-		return this.ax;
+		return this.p1.xProperty();
 	}
 
 	@Pure
 	@Override
 	public double getY1() {
-		return this.ay == null ? 0 : this.ay.get();
+		return this.p1.getY();
 	}
 
 	/** Replies the property that is the y coordinate of the first segment point.
@@ -190,16 +220,13 @@ public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
 	 */
 	@Pure
 	public DoubleProperty y1Property() {
-		if (this.ay == null) {
-			this.ay = new SimpleDoubleProperty(this, MathFXAttributeNames.Y1);
-		}
-		return this.ay;
+		return this.p1.yProperty();
 	}
 
 	@Pure
 	@Override
 	public double getX2() {
-		return this.bx == null ? 0 : this.bx.get();
+		return this.p2.getX();
 	}
 
 	/** Replies the property that is the x coordinate of the second segment point.
@@ -208,16 +235,13 @@ public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
 	 */
 	@Pure
 	public DoubleProperty x2Property() {
-		if (this.bx == null) {
-			this.bx = new SimpleDoubleProperty(this, MathFXAttributeNames.X2);
-		}
-		return this.bx;
+		return this.p2.xProperty();
 	}
 
 	@Pure
 	@Override
 	public double getY2() {
-		return this.by == null ? 0 : this.by.get();
+		return this.p2.getY();
 	}
 
 	/** Replies the property that is the x coordinate of the second segment point.
@@ -226,29 +250,63 @@ public class Segment2dfx extends AbstractShape2dfx<Segment2dfx>
 	 */
 	@Pure
 	public DoubleProperty y2Property() {
-		if (this.by == null) {
-			this.by = new SimpleDoubleProperty(this, MathFXAttributeNames.Y2);
-		}
-		return this.by;
+		return this.p2.yProperty();
 	}
 
 	@Override
 	public Point2dfx getP1() {
-		return getGeomFactory().newPoint(this.ax, this.ay);
+		return this.p1;
+	}
+
+	@Override
+	public void setP1(double x, double y) {
+	    this.p1.set(x, y);
+	}
+
+	@Override
+	public void setP1(Point2D<?, ?> pt) {
+	    this.p1.setX(pt.getX());
+	    this.p1.setY(pt.getY());
+	}
+
+	/** Set the oriented point as the first point of this Segment2dfx to preserve length.
+	 * @param pt the point.
+	 */
+	public void setP1(Point2dfx pt) {
+	    this.p1 = pt;
 	}
 
 	@Override
 	public Point2dfx getP2() {
-		return getGeomFactory().newPoint(this.bx, this.by);
+		return this.p2;
+	}
+
+	@Override
+	public void setP2(double x, double y) {
+		this.p2.set(x, y);
+	}
+
+	@Override
+	public void setP2(Point2D<?, ?> pt) {
+		this.p2.setX(pt.getX());
+		this.p2.setY(pt.getY());
+	}
+
+	/** Set the oriented point as the second point of this Segment2dfx to preserve length.
+	 * @param pt the point.
+	 */
+	public void setP2(Point2dfx pt) {
+	    this.p2 = pt;
 	}
 
 	@Override
 	public ObjectProperty<Rectangle2dfx> boundingBoxProperty() {
 		if (this.boundingBox == null) {
 			this.boundingBox = new SimpleObjectProperty<>(this, MathFXAttributeNames.BOUNDING_BOX);
-			this.boundingBox.bind(Bindings.createObjectBinding(() -> toBoundingBox(),
-					x1Property(), y1Property(),
-					x2Property(), y2Property()));
+			this.boundingBox.bind(Bindings.createObjectBinding(() -> {
+			    return toBoundingBox();
+			}, x1Property(), y1Property(),
+			   x2Property(), y2Property()));
 		}
 		return this.boundingBox;
 	}
