+++ advanced/mathfx/src/main/java/org/arakhne/afc/math/geometry/d3/ifx/Segment3ifx.java
@@ -31,12 +31,11 @@ import org.arakhne.afc.math.geometry.MathFXAttributeNames;
 import org.arakhne.afc.math.geometry.d3.Point3D;
 import org.arakhne.afc.math.geometry.d3.Transform3D;
 import org.arakhne.afc.math.geometry.d3.ai.Segment3ai;
-import org.arakhne.afc.vmutil.asserts.AssertMessages;
 
 /** A 3D segment/line with 3 integer FX properties.
  *
- * @author $Author: sgalland$
  * @author $Author: tpiotrow$
+ * @author $Author: sgalland$
  * @version $FullVersion$
  * @mavengroupid $GroupId$
  * @mavenartifactid $ArtifactId$
@@ -47,17 +46,9 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 
 	private static final long serialVersionUID = -1406743357357708790L;
 
-	private IntegerProperty ax;
+	private Point3ifx p1 = new Point3ifx();
 
-	private IntegerProperty ay;
-
-	private IntegerProperty az;
-
-	private IntegerProperty bx;
-
-	private IntegerProperty by;
-
-	private IntegerProperty bz;
+	private Point3ifx p2 = new Point3ifx();
 
 	/** Construct an empty segment.
      */
@@ -66,18 +57,35 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	}
 
 	/** Construct a segment with the two given points.
+	 * @param p1 first point.
+	 * @param p2 second point.
+	 */
+	public Segment3ifx(Point3D<?, ?> p1, Point3D<?, ?> p2) {
+	    this(p1.ix(), p1.iy(), p1.iz(), p2.ix(), p2.iy(), p2.iz());
+	}
+
+	/** Construct a segment with the two given points.
      * @param p1 first point.
      * @param p2 second point.
      */
-	public Segment3ifx(Point3D<?, ?> p1, Point3D<?, ?> p2) {
-		this(p1.ix(), p1.iy(), p1.iz(), p2.ix(), p2.iy(), p2.iz());
+	public Segment3ifx(Point3ifx p1, Point3ifx p2) {
+	    this.p1 = p1;
+	    this.p2 = p2;
+	}
+
+	/** Construct by copy.
+	 * @param segment the segment to copy.
+	 */
+	public Segment3ifx(Segment3ai<?, ?, ?, ?, ?, ?> segment) {
+	    this(segment.getX1(), segment.getY1(), segment.getZ1(), segment.getX2(), segment.getY2(), segment.getZ2());
 	}
 
 	/** Construct by copy.
      * @param segment the segment to copy.
      */
-	public Segment3ifx(Segment3ai<?, ?, ?, ?, ?, ?> segment) {
-		this(segment.getX1(), segment.getY1(), segment.getZ1(), segment.getX2(), segment.getY2(), segment.getZ2());
+	public Segment3ifx(Segment3ifx segment) {
+	    this.p1 = segment.p1;
+	    this.p2 = segment.p2;
 	}
 
 	/** Construct a segment with the two given points.
@@ -95,29 +103,13 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	@Override
 	public Segment3ifx clone() {
 		final Segment3ifx clone = super.clone();
-		if (clone.ax != null) {
-			clone.ax = null;
-			clone.x1Property().set(getX1());
+		if (clone.p1 != null) {
+			clone.p1 = null;
+			clone.p1 = this.p1.clone();
 		}
-		if (clone.ay != null) {
-			clone.ay = null;
-			clone.y1Property().set(getY1());
-		}
-		if (clone.az != null) {
-			clone.az = null;
-			clone.z1Property().set(getZ1());
-		}
-		if (clone.bx != null) {
-			clone.bx = null;
-			clone.x2Property().set(getX2());
-		}
-		if (clone.by != null) {
-			clone.by = null;
-			clone.y2Property().set(getY2());
-		}
-		if (clone.bz != null) {
-			clone.bz = null;
-			clone.z2Property().set(getZ2());
+		if (clone.p2 != null) {
+		    clone.p2 = null;
+			clone.p2 = this.p2.clone();
 		}
 		return clone;
 	}
@@ -138,7 +130,9 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	@Pure
 	@Override
 	public Shape3ifx<?> createTransformedShape(Transform3D transform) {
-		assert transform != null : AssertMessages.notNullParameter();
+        if (transform == null || transform.isIdentity()) {
+            return clone();
+        }
 		final Point3ifx point = getGeomFactory().newPoint(getX1(), getY1(), getZ1());
 		transform.transform(point);
 		final int x1 = point.ix();
@@ -192,7 +186,7 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	@Pure
 	@Override
 	public int getX1() {
-		return this.ax == null ? 0 : this.ax.get();
+		return this.p1.x == null ? 0 : this.p1.x.get();
 	}
 
 	/** Replies the property that is the x coordinate of the first segment point.
@@ -201,16 +195,16 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	 */
 	@Pure
 	public IntegerProperty x1Property() {
-		if (this.ax == null) {
-			this.ax = new SimpleIntegerProperty(this, MathFXAttributeNames.X1);
+		if (this.p1.x == null) {
+			this.p1.x = new SimpleIntegerProperty(this, MathFXAttributeNames.X1);
 		}
-		return this.ax;
+		return this.p1.x;
 	}
 
 	@Pure
 	@Override
 	public int getY1() {
-		return this.ay == null ? 0 : this.ay.get();
+		return this.p1.y == null ? 0 : this.p1.y.get();
 	}
 
 	/** Replies the property that is the y coordinate of the first segment point.
@@ -219,16 +213,16 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	 */
 	@Pure
 	public IntegerProperty y1Property() {
-		if (this.ay == null) {
-			this.ay = new SimpleIntegerProperty(this, MathFXAttributeNames.Y1);
+		if (this.p1.y == null) {
+			this.p1.y = new SimpleIntegerProperty(this, MathFXAttributeNames.Y1);
 		}
-		return this.ay;
+		return this.p1.y;
 	}
 
 	@Pure
 	@Override
 	public int getZ1() {
-		return this.az == null ? 0 : this.az.get();
+		return this.p1.z == null ? 0 : this.p1.z.get();
 	}
 
 	/** Replies the property that is the z coordinate of the first segment point.
@@ -237,16 +231,16 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	 */
 	@Pure
 	public IntegerProperty z1Property() {
-		if (this.az == null) {
-			this.az = new SimpleIntegerProperty(this, MathFXAttributeNames.Z1);
+		if (this.p1.z == null) {
+			this.p1.z = new SimpleIntegerProperty(this, MathFXAttributeNames.Z1);
 		}
-		return this.az;
+		return this.p1.z;
 	}
 
 	@Pure
 	@Override
 	public int getX2() {
-		return this.bx == null ? 0 : this.bx.get();
+		return this.p2.x == null ? 0 : this.p2.x.get();
 	}
 
 	/** Replies the property that is the x coordinate of the second segment point.
@@ -255,16 +249,16 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	 */
 	@Pure
 	public IntegerProperty x2Property() {
-		if (this.bx == null) {
-			this.bx = new SimpleIntegerProperty(this, MathFXAttributeNames.X2);
+		if (this.p2.x == null) {
+			this.p2.x = new SimpleIntegerProperty(this, MathFXAttributeNames.X2);
 		}
-		return this.bx;
+		return this.p2.x;
 	}
 
 	@Pure
 	@Override
 	public int getY2() {
-		return this.by == null ? 0 : this.by.get();
+		return this.p2.y == null ? 0 : this.p2.y.get();
 	}
 
 	/** Replies the property that is the y coordinate of the second segment point.
@@ -273,16 +267,16 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	 */
 	@Pure
 	public IntegerProperty y2Property() {
-		if (this.by == null) {
-			this.by = new SimpleIntegerProperty(this, MathFXAttributeNames.Y2);
+		if (this.p2.y == null) {
+			this.p2.y = new SimpleIntegerProperty(this, MathFXAttributeNames.Y2);
 		}
-		return this.by;
+		return this.p2.y;
 	}
 
 	@Pure
 	@Override
 	public int getZ2() {
-		return this.bz == null ? 0 : this.bz.get();
+		return this.p2.z == null ? 0 : this.p2.z.get();
 	}
 
 	/** Replies the property that is the z coordinate of the second segment point.
@@ -291,20 +285,60 @@ public class Segment3ifx extends AbstractShape3ifx<Segment3ifx>
 	 */
 	@Pure
 	public IntegerProperty z2Property() {
-		if (this.bz == null) {
-			this.bz = new SimpleIntegerProperty(this, MathFXAttributeNames.Z2);
+		if (this.p2.z == null) {
+			this.p2.z = new SimpleIntegerProperty(this, MathFXAttributeNames.Z2);
 		}
-		return this.bz;
+		return this.p2.z;
 	}
 
 	@Override
 	public Point3ifx getP1() {
-		return getGeomFactory().newPoint(this.ax, this.ay, this.az);
+		return this.p1;
+	}
+
+	@Override
+	public void setP1(int x, int y, int z) {
+	    this.p1.set(x, y, z);
+	}
+
+	@Override
+	public void setP1(Point3D<?, ?> point) {
+	    this.p1.setX(point.getX());
+	    this.p1.setY(point.getY());
+	    this.p1.setZ(point.getZ());
+	}
+
+	/** Set the point as the first point of this segment.
+	 *
+	 * @param point the point to set.
+	 */
+	public void setP1(Point3ifx point) {
+	    this.p1 = point;
 	}
 
 	@Override
 	public Point3ifx getP2() {
-		return getGeomFactory().newPoint(this.bx, this.by, this.bz);
+		return this.p2;
+	}
+
+	@Override
+	public void setP2(int x, int y, int z) {
+	    this.p2.set(x, y, z);
+	}
+
+	@Override
+	public void setP2(Point3D<?, ?> point) {
+	    this.p2.setX(point.getX());
+	    this.p2.setY(point.getY());
+	    this.p2.setZ(point.getZ());
+	}
+
+	/** Set the point as the second point of this segment.
+	 *
+	 * @param point the point to set.
+	 */
+	public void setP2(Point3ifx point) {
+	    this.p2 = point;
 	}
 
 	@Override
