+++ advanced/mathfx/src/main/java/org/arakhne/afc/math/geometry/d3/ifx/Segment3ifxOld.java
@@ -31,61 +31,53 @@ import org.arakhne.afc.math.geometry.MathFXAttributeNames;
 import org.arakhne.afc.math.geometry.d3.Point3D;
 import org.arakhne.afc.math.geometry.d3.Transform3D;
 import org.arakhne.afc.math.geometry.d3.ai.Segment3ai;
+import org.arakhne.afc.vmutil.asserts.AssertMessages;
 
 /** A 3D segment/line with 3 integer FX properties.
  *
- * @author $Author: tpiotrow$
  * @author $Author: sgalland$
+ * @author $Author: tpiotrow$
  * @version $FullVersion$
  * @mavengroupid $GroupId$
  * @mavenartifactid $ArtifactId$
  * @since 13.0
  */
-public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
-	implements Segment3ai<Shape3ifx<?>, SegmentPoint3ifx, PathElement3ifx, Point3ifx, Vector3ifx, RectangularPrism3ifx> {
+public class Segment3ifxOld extends AbstractShape3ifx<Segment3ifx>
+	implements Segment3ai<Shape3ifx<?>, Segment3ifx, PathElement3ifx, Point3ifx, Vector3ifx, RectangularPrism3ifx> {
 
 	private static final long serialVersionUID = -1406743357357708790L;
 
-	private Point3ifx p1 = new Point3ifx();
+	private IntegerProperty ax;
 
-	private Point3ifx p2 = new Point3ifx();
+	private IntegerProperty ay;
+
+	private IntegerProperty az;
+
+	private IntegerProperty bx;
+
+	private IntegerProperty by;
+
+	private IntegerProperty bz;
 
 	/** Construct an empty segment.
      */
-	public SegmentPoint3ifx() {
+	public Segment3ifxOld() {
 		//
-	}
-
-	/** Construct a segment with the two given points.
-	 * @param p1 first point.
-	 * @param p2 second point.
-	 */
-	public SegmentPoint3ifx(Point3D<?, ?> p1, Point3D<?, ?> p2) {
-	    this(p1.ix(), p1.iy(), p1.iz(), p2.ix(), p2.iy(), p2.iz());
 	}
 
 	/** Construct a segment with the two given points.
      * @param p1 first point.
      * @param p2 second point.
      */
-	public SegmentPoint3ifx(Point3ifx p1, Point3ifx p2) {
-	    this.p1 = p1;
-	    this.p2 = p2;
-	}
-
-	/** Construct by copy.
-	 * @param segment the segment to copy.
-	 */
-	public SegmentPoint3ifx(Segment3ai<?, ?, ?, ?, ?, ?> segment) {
-	    this(segment.getX1(), segment.getY1(), segment.getZ1(), segment.getX2(), segment.getY2(), segment.getZ2());
+	public Segment3ifxOld(Point3D<?, ?> p1, Point3D<?, ?> p2) {
+		this(p1.ix(), p1.iy(), p1.iz(), p2.ix(), p2.iy(), p2.iz());
 	}
 
 	/** Construct by copy.
      * @param segment the segment to copy.
      */
-	public SegmentPoint3ifx(SegmentPoint3ifx segment) {
-	    this.p1 = segment.p1;
-	    this.p2 = segment.p2;
+	public Segment3ifxOld(Segment3ai<?, ?, ?, ?, ?, ?> segment) {
+		this(segment.getX1(), segment.getY1(), segment.getZ1(), segment.getX2(), segment.getY2(), segment.getZ2());
 	}
 
 	/** Construct a segment with the two given points.
@@ -96,23 +88,37 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
      * @param y2 y coordinate of the second point.
      * @param z2 z coordinate of the second point.
      */
-	public SegmentPoint3ifx(int x1, int y1, int z1, int x2, int y2, int z2) {
+	public Segment3ifxOld(int x1, int y1, int z1, int x2, int y2, int z2) {
 		set(x1, y1, z1, x2, y2, z2);
 	}
 
 	@Override
-	public SegmentPoint3ifx clone() {
-		final SegmentPoint3ifx clone = super.clone();
-		if (clone.p1 != null) {
-			clone.x1Property().set(getX1());
-			clone.y1Property().set(getY1());
-			clone.z1Property().set(getZ1());
-		}
-		if (clone.p2 != null) {
-			clone.x2Property().set(getX2());
-			clone.y2Property().set(getY2());
-			clone.z2Property().set(getZ2());
-		}
+	public Segment3ifx clone() {
+		final Segment3ifx clone = super.clone();
+		//		if (clone.ax != null) {
+		//			clone.ax = null;
+		//			clone.x1Property().set(getX1());
+		//		}
+		//		if (clone.ay != null) {
+		//			clone.ay = null;
+		//			clone.y1Property().set(getY1());
+		//		}
+		//		if (clone.az != null) {
+		//			clone.az = null;
+		//			clone.z1Property().set(getZ1());
+		//		}
+		//		if (clone.bx != null) {
+		//			clone.bx = null;
+		//			clone.x2Property().set(getX2());
+		//		}
+		//		if (clone.by != null) {
+		//			clone.by = null;
+		//			clone.y2Property().set(getY2());
+		//		}
+		//		if (clone.bz != null) {
+		//			clone.bz = null;
+		//			clone.z2Property().set(getZ2());
+		//		}
 		return clone;
 	}
 
@@ -132,9 +138,7 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
 	@Pure
 	@Override
 	public Shape3ifx<?> createTransformedShape(Transform3D transform) {
-        if (transform == null || transform.isIdentity()) {
-            return clone();
-        }
+		assert transform != null : AssertMessages.notNullParameter();
 		final Point3ifx point = getGeomFactory().newPoint(getX1(), getY1(), getZ1());
 		transform.transform(point);
 		final int x1 = point.ix();
@@ -188,7 +192,7 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
 	@Pure
 	@Override
 	public int getX1() {
-		return this.p1.x == null ? 0 : this.p1.x.get();
+		return this.ax == null ? 0 : this.ax.get();
 	}
 
 	/** Replies the property that is the x coordinate of the first segment point.
@@ -197,16 +201,16 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
 	 */
 	@Pure
 	public IntegerProperty x1Property() {
-		if (this.p1.x == null) {
-			this.p1.x = new SimpleIntegerProperty(this, MathFXAttributeNames.X1);
+		if (this.ax == null) {
+			this.ax = new SimpleIntegerProperty(this, MathFXAttributeNames.X1);
 		}
-		return this.p1.x;
+		return this.ax;
 	}
 
 	@Pure
 	@Override
 	public int getY1() {
-		return this.p1.y == null ? 0 : this.p1.y.get();
+		return this.ay == null ? 0 : this.ay.get();
 	}
 
 	/** Replies the property that is the y coordinate of the first segment point.
@@ -215,16 +219,16 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
 	 */
 	@Pure
 	public IntegerProperty y1Property() {
-		if (this.p1.y == null) {
-			this.p1.y = new SimpleIntegerProperty(this, MathFXAttributeNames.Y1);
+		if (this.ay == null) {
+			this.ay = new SimpleIntegerProperty(this, MathFXAttributeNames.Y1);
 		}
-		return this.p1.y;
+		return this.ay;
 	}
 
 	@Pure
 	@Override
 	public int getZ1() {
-		return this.p1.z == null ? 0 : this.p1.z.get();
+		return this.az == null ? 0 : this.az.get();
 	}
 
 	/** Replies the property that is the z coordinate of the first segment point.
@@ -233,16 +237,16 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
 	 */
 	@Pure
 	public IntegerProperty z1Property() {
-		if (this.p1.z == null) {
-			this.p1.z = new SimpleIntegerProperty(this, MathFXAttributeNames.Z1);
+		if (this.az == null) {
+			this.az = new SimpleIntegerProperty(this, MathFXAttributeNames.Z1);
 		}
-		return this.p1.z;
+		return this.az;
 	}
 
 	@Pure
 	@Override
 	public int getX2() {
-		return this.p2.x == null ? 0 : this.p2.x.get();
+		return this.bx == null ? 0 : this.bx.get();
 	}
 
 	/** Replies the property that is the x coordinate of the second segment point.
@@ -251,16 +255,16 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
 	 */
 	@Pure
 	public IntegerProperty x2Property() {
-		if (this.p2.x == null) {
-			this.p2.x = new SimpleIntegerProperty(this, MathFXAttributeNames.X2);
+		if (this.bx == null) {
+			this.bx = new SimpleIntegerProperty(this, MathFXAttributeNames.X2);
 		}
-		return this.p2.x;
+		return this.bx;
 	}
 
 	@Pure
 	@Override
 	public int getY2() {
-		return this.p2.y == null ? 0 : this.p2.y.get();
+		return this.by == null ? 0 : this.by.get();
 	}
 
 	/** Replies the property that is the y coordinate of the second segment point.
@@ -269,16 +273,16 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
 	 */
 	@Pure
 	public IntegerProperty y2Property() {
-		if (this.p2.y == null) {
-			this.p2.y = new SimpleIntegerProperty(this, MathFXAttributeNames.Y2);
+		if (this.by == null) {
+			this.by = new SimpleIntegerProperty(this, MathFXAttributeNames.Y2);
 		}
-		return this.p2.y;
+		return this.by;
 	}
 
 	@Pure
 	@Override
 	public int getZ2() {
-		return this.p2.z == null ? 0 : this.p2.z.get();
+		return this.bz == null ? 0 : this.bz.get();
 	}
 
 	/** Replies the property that is the z coordinate of the second segment point.
@@ -287,60 +291,20 @@ public class SegmentPoint3ifx extends AbstractShape3ifx<SegmentPoint3ifx>
 	 */
 	@Pure
 	public IntegerProperty z2Property() {
-		if (this.p2.z == null) {
-			this.p2.z = new SimpleIntegerProperty(this, MathFXAttributeNames.Z2);
+		if (this.bz == null) {
+			this.bz = new SimpleIntegerProperty(this, MathFXAttributeNames.Z2);
 		}
-		return this.p2.z;
+		return this.bz;
 	}
 
 	@Override
 	public Point3ifx getP1() {
-		return this.p1;
-	}
-
-	@Override
-	public void setP1(int x, int y, int z) {
-	    this.p1.set(x, y, z);
-	}
-
-	@Override
-	public void setP1(Point3D<?, ?> point) {
-	    this.p1.setX(point.getX());
-	    this.p1.setY(point.getY());
-	    this.p1.setZ(point.getZ());
-	}
-
-	/** Set the point as the first point of this segment.
-	 *
-	 * @param point the point to set.
-	 */
-	public void setP1(Point3ifx point) {
-	    this.p1 = point;
+		return getGeomFactory().newPoint(this.ax, this.ay, this.az);
 	}
 
 	@Override
 	public Point3ifx getP2() {
-		return this.p2;
-	}
-
-	@Override
-	public void setP2(int x, int y, int z) {
-	    this.p2.set(x, y, z);
-	}
-
-	@Override
-	public void setP2(Point3D<?, ?> point) {
-	    this.p2.setX(point.getX());
-	    this.p2.setY(point.getY());
-	    this.p2.setZ(point.getZ());
-	}
-
-	/** Set the point as the second point of this segment.
-	 *
-	 * @param point the point to set.
-	 */
-	public void setP2(Point3ifx point) {
-	    this.p2 = point;
+		return getGeomFactory().newPoint(this.bx, this.by, this.bz);
 	}
 
 	@Override
